Split the zebra_state::service::read module
